requirements.txt
flask>=3.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
